refactor(gsm): move the modem transport into connections/gsm/
Closes item 8 of #611. SerialCommunication was a parallel hand-rolled
reimplementation of the Connection abstraction living in interfaces/, and
it shared its name with connections/serial/connection.py's panel transport
-- the same non-substitutable collision item 7 fixed for the protocols.

- New paradox/connections/gsm/ package holding GsmSerialProtocol and
  GsmSerialConnection, mirroring ip/, serial/ and prt3/.
- GsmSerialConnection extends Connection for lifecycle only: the
  protocol-aware connected property, _protocol plumbing, and close(), which
  the GSM transport had no equivalent of. on_message stays overridden to
  keep the queue and recv_callback dispatch, because AT command/response
  does not fit Connection's parsed-message handler registry.
- The request/response method is renamed send_command(). Keeping it as
  write() would have collided with Connection.write(), which is a
  synchronous fire-and-forget and would leave send_message's coroutine
  un-awaited.
- GSMTextInterface.stop() now closes the port instead of logging a TODO.

Bugs fixed in the moved code, all of which connections/serial/connection.py
already got right:

- on_connection and on_connection_loss set connected_future unguarded, so a
  modem drop after a successful connect raised InvalidStateError.
- The open timeout handle was never cancelled, and the constructor's timeout
  argument was accepted and then ignored in favour of a hardcoded 5.
- send_command's timeout argument was ignored in favour of a hardcoded 5.
- The port check was os.path.exists, which misses the far more common
  permission failure that os.access(R_OK|W_OK) catches.
- connect() logged "Could not connect" via logger.exception outside any
  except block, emitting a bogus "NoneType: None" traceback.
