+++ paradox/connections/gsm/__init__.py
@@ -1,0 +1,1 @@
+# GSM connection package — AT-command serial transport for a GSM modem.
